+++ .idea/gradle.xml
@@ -12,7 +12,10 @@
           <set>
             <option value="$PROJECT_DIR$" />
             <option value="$PROJECT_DIR$/app" />
+            <option value="$PROJECT_DIR$/data" />
+            <option value="$PROJECT_DIR$/local" />
             <option value="$PROJECT_DIR$/presentation" />
+            <option value="$PROJECT_DIR$/remote" />
           </set>
         </option>
         <option name="resolveModulePerSourceSet" value="false" />
