+++ data/build.gradle
@@ -15,6 +15,7 @@ android {
         versionName "1.0"
 
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
+        consumerProguardFiles "consumer-rules.pro"
     }
 
     buildTypes {
@@ -22,11 +23,6 @@ android {
             minifyEnabled false
             proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
         }
-    }
-    buildFeatures{
-//        viewBinding = true
-        dataBinding = true
-
     }
     compileOptions {
         sourceCompatibility JavaVersion.VERSION_1_8
@@ -38,10 +34,7 @@ android {
 }
 
 dependencies {
-    implementation project(':data')
-    implementation project(':local')
-    implementation project(':remote')
-    
+
     /* Retrofit2 */
     implementation 'com.squareup.retrofit2:retrofit:2.9.0'
     implementation 'com.squareup.retrofit2:converter-gson:2.9.0'
@@ -68,12 +61,4 @@ dependencies {
     androidTestImplementation 'androidx.test.ext:junit:1.1.2'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
     kapt "androidx.room:room-compiler:2.4.3"
-
-    // navigation
-    implementation 'androidx.navigation:navigation-fragment-ktx:2.5.3'
-    implementation 'androidx.navigation:navigation-ui-ktx:2.5.3'
-
-    // For BottomNavigationView from Material Components
-    implementation 'com.google.android.material:material:1.1.0'
-    implementation "com.jakewharton.timber:timber:4.7.1"
 }
