+++ data/src/main/java/com/example/data/model/DataNewsModel.kt
@@ -1,0 +1,31 @@
+/**
+ * Copyright (C) 2023. ExodusEnt Corp. All rights reserved.
+ * You must have prior written permission to read this file.
+ * @author Kim min gyu <mingue0605@myloveidol.com>
+ * Description:
+ *
+ * */
+
+package com.example.data.model
+
+import android.os.Parcelable
+import androidx.room.ColumnInfo
+import androidx.room.Entity
+import androidx.room.PrimaryKey
+import kotlinx.android.parcel.Parcelize
+
+data class DataNewsModel(
+    var articles: List<Articles> = listOf()
+)
+
+@Entity(tableName = "articles")
+@Parcelize
+data class Articles(
+    @PrimaryKey @ColumnInfo(name = "url") var url: String,
+    @ColumnInfo(name = "author") var author: String? = null,
+    @ColumnInfo(name = "title") var title: String,
+    @ColumnInfo(name = "description") var description: String? = null,
+    @ColumnInfo(name = "urlToImage") var urlToImage: String? = null,
+    @ColumnInfo(name = "publishedAt")  var publishedAt: String? = null,
+    @ColumnInfo(name = "content") var content: String? = null,
+) : Parcelable
