+++ data/src/main/java/com/example/data/repository/NewsRepositoryImpl.kt
@@ -1,0 +1,36 @@
+/**
+ * Copyright (C) 2023. ExodusEnt Corp. All rights reserved.
+ * You must have prior written permission to read this file.
+ * @author Kim min gyu <mingue0605@myloveidol.com>
+ * Description:
+ *
+ * */
+
+package com.example.data.repository
+
+import com.example.data.local.LocalDataSource
+import com.example.data.model.Articles
+import com.example.data.model.DataNewsModel
+import com.example.data.remote.RemoteDataSource
+import kotlinx.coroutines.flow.Flow
+
+class NewsRepositoryImpl(
+    private val remoteDataSource: RemoteDataSource,
+    private val localDataSource: LocalDataSource
+) : NewsRepository {
+    override suspend fun getNews(country: String, category: String?): Flow<DataNewsModel> = remoteDataSource.getNews(country, category)
+
+    override suspend fun getAll(): Flow<List<Articles>> = localDataSource.getAll()
+
+    override fun insert(items: Articles, callback: () -> Unit) {
+        localDataSource.insert(items){
+            callback()
+        }
+    }
+
+    override fun deleteArticle(url: String, callback: () -> Unit) {
+        localDataSource.deleteArticle(url){
+            callback()
+        }
+    }
+}
