+++ local/src/main/java/com/example/local/dataSource/LocalDataSourceImpl.kt
@@ -1,0 +1,42 @@
+/**
+ * Copyright (C) 2023. ExodusEnt Corp. All rights reserved.
+ * You must have prior written permission to read this file.
+ * @author Kim min gyu <mingue0605@myloveidol.com>
+ * Description:
+ *
+ * */
+
+package com.example.local.dataSource
+
+import com.example.data.local.LocalDataSource
+import com.example.data.model.Articles
+import com.example.local.Room.AppDB
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.flow
+import kotlinx.coroutines.flow.flowOn
+
+
+/**
+ * @see
+ * */
+
+class LocalDataSourceImpl(
+    private val appDB: AppDB
+) : LocalDataSource {
+    override suspend fun getAll(): Flow<List<Articles>> = flow {
+        val result = appDB.newsDao().getAll()
+        emit(result)
+    }.flowOn(Dispatchers.IO)
+
+    override fun insert(items: Articles, callback: () -> Unit) {
+        appDB.newsDao().insert(items)
+        callback()
+    }
+
+    override fun deleteArticle(url: String, callback: () -> Unit) {
+        appDB.newsDao().deleteArticle(url)
+        callback()
+    }
+
+}
