+++ local/src/main/java/com/example/local/mapper/LocalMapper.kt
@@ -1,0 +1,15 @@
+/**
+ * Copyright (C) 2023. ExodusEnt Corp. All rights reserved.
+ * You must have prior written permission to read this file.
+ * @author Kim min gyu <mingue0605@myloveidol.com>
+ * Description:
+ *
+ * */
+
+package com.example.local.mapper
+
+
+interface LocalMapper<T, E> {
+    fun T.toData(): E
+    fun E.fromData(): T
+}
