+++ presentation/src/main/java/com/example/presentation/Fragment/CategoryNewsFragment.kt
@@ -1,35 +1,23 @@
 package com.example.presentation.Fragment
 
 import android.os.Bundle
-import android.util.Log
-import android.view.LayoutInflater
 import android.view.View
-import android.view.ViewGroup
-import androidx.core.os.bundleOf
-import androidx.fragment.app.Fragment
 import androidx.navigation.NavController
 import androidx.navigation.fragment.NavHostFragment
 import androidx.recyclerview.widget.LinearLayoutManager
+import com.example.data.model.Articles
+import com.example.data.repository.NewsRepository
+import com.example.data.repository.NewsRepositoryImpl
 import com.example.presentation.*
 import com.example.presentation.Adapter.NewsListAdapter
-import com.example.presentation.Room.AppDB
+import com.example.local.Room.AppDB
 import com.example.presentation.databinding.FragmentCategoryNewsBinding
-import com.example.presentation.datasource.local.LocalDataSourceImpl
-import com.example.presentation.datasource.remote.RemoteDataSourceImpl
-import com.example.presentation.repository.NewsRepository
-import com.example.presentation.repository.NewsRepositoryImpl
-import com.example.presentation.retrofit.ApiService
-import com.example.presentation.retrofit.RetrofitHelper
+import com.example.local.dataSource.LocalDataSourceImpl
+import com.example.remote.dataSource.RemoteDataSourceImpl
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
-import retrofit2.Call
-import retrofit2.Callback
-import retrofit2.Response
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-import timber.log.Timber
 
 class CategoryNewsFragment : BaseFragment<FragmentCategoryNewsBinding>(R.layout.fragment_category_news), NewsListAdapter.OnClickListener{
 
@@ -40,7 +28,9 @@ class CategoryNewsFragment : BaseFragment<FragmentCategoryNewsBinding>(R.layout.
 
     private val categoryNewsFragmentRepository : NewsRepository by lazy{
         val remoteDataSourceImpl = RemoteDataSourceImpl()
-        val localDataSourceImpl = LocalDataSourceImpl(AppDB.getInstance(requireActivity()))
+        val localDataSourceImpl = LocalDataSourceImpl(
+            AppDB.getInstance(requireActivity())
+        )
 
         NewsRepositoryImpl(remoteDataSourceImpl, localDataSourceImpl)
     }
