--- presentation/src/main/java/com/example/presentation/Room/AppDB.kt
@@ -1,30 +1,0 @@
-package com.example.presentation.Room
-
-import android.content.Context
-import androidx.room.Database
-import androidx.room.Room
-import androidx.room.RoomDatabase
-import com.example.presentation.Articles
-
-@Database(entities = [Articles::class], version = 1, exportSchema = false)
-abstract class AppDB : RoomDatabase() {
-    abstract fun newsDao() : NewsDao
-
-    companion object {
-        private var INSTANCE: AppDB? = null
-
-        //Singleton pattern
-        fun getInstance(context: Context): AppDB {
-            if (INSTANCE == null) {
-                //synchronized : 중복 방지
-                synchronized(AppDB::class) {
-                    INSTANCE = Room.databaseBuilder(context.applicationContext,
-                        AppDB::class.java, "recipe.db")
-                        .fallbackToDestructiveMigration()
-                        .build()
-                }
-            }
-            return INSTANCE!!
-        }
-    }
-}
