+++ presentation/src/main/java/com/example/presentation/model/NewsModel.kt
@@ -1,0 +1,10 @@
+/**
+ * Copyright (C) 2023. ExodusEnt Corp. All rights reserved.
+ * You must have prior written permission to read this file.
+ * @author Kim min gyu <mingue0605@myloveidol.com>
+ * Description:
+ *
+ * */
+
+package com.example.presentation.model
+
