+++ remote/build.gradle
@@ -39,9 +39,7 @@ android {
 
 dependencies {
     implementation project(':data')
-    implementation project(':local')
-    implementation project(':remote')
-    
+
     /* Retrofit2 */
     implementation 'com.squareup.retrofit2:retrofit:2.9.0'
     implementation 'com.squareup.retrofit2:converter-gson:2.9.0'
@@ -68,12 +66,4 @@ dependencies {
     androidTestImplementation 'androidx.test.ext:junit:1.1.2'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
     kapt "androidx.room:room-compiler:2.4.3"
-
-    // navigation
-    implementation 'androidx.navigation:navigation-fragment-ktx:2.5.3'
-    implementation 'androidx.navigation:navigation-ui-ktx:2.5.3'
-
-    // For BottomNavigationView from Material Components
-    implementation 'com.google.android.material:material:1.1.0'
-    implementation "com.jakewharton.timber:timber:4.7.1"
 }
